project restructure and daily page styles made
