--- assets/tasks/tasks.css
@@ -1,75 +1,0 @@
-.weekly-lists{
-    display: flex;
-    width: 100vw;
-    height: 84vh;
-    overflow-x: scroll;
-    overflow-y: hidden;
-    scroll-snap-type: x mandatory;
-    box-sizing: border-box;
-}
-
-.page {
-    height: 100%;
-    flex-shrink: 0;
-    padding: 8rem 0;
-    box-sizing: border-box;
-}
-
-#rewards {
-    display: flex;
-    width: 70vw;
-    scroll-snap-align: Start;
-    gap: 2vw;
-    padding-left: 2vw;
-    padding-right: 2vw;
-}
-
-.reward {
-    background-color: var(--accent1_dark);
-    flex-grow: 1;
-}
-
-.checks {
-    background-color: var(--accent1_dark);
-    width: 28vw;
-}
-
-#calendar {
-    width: 36vw;
-    scroll-snap-align: Center;
-}
-
-#calendar.small {
-    width: 28vw;
-}
-
-.calendar {
-    background-color: var(--accent1_dark);
-    width: 100%;
-    height: 100%;
-}
-
-#todays {
-    width: 30vw;
-    padding-left: 2vw;
-}
-
-#weekly {
-    display: flex;
-    gap: 2vw;
-
-    scroll-snap-align: End;
-
-    width: 62vw;
-    padding-left: 2vw;
-    padding-right: 2vw;
-}
-
-#tasks {
-    display: flex;
-
-    scroll-snap-align: End;
-
-    width: 30vw;
-    padding-right: 2vw;
-}
