+++ src/components/todo/rewards/mod.rs
@@ -1,0 +1,10 @@
+use dioxus::prelude::*;
+
+#[component]
+pub fn Checks() -> Element {
+    rsx!{
+        div {
+            class: "checks",
+        }
+    }
+}
